implémentation de book et chapter

+++ src/main/java/fr/scriptoria/api/repositories/mySqlRepositories/ProjectRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import fr.scriptoria.api.models.entities.ProjectEntity;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<ProjectEntity, Integer> {
+public interface ProjectRepository extends JpaRepository<ProjectEntity, Long> {
     
 }
